moved redis store to common
